packages/drawer/pyproject.toml
[project]
name = "drawer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gpiozero>=2.0.1",
    "imagen",
    "lgpio>=0.2.2.0",
    "pillow>=11.0.0",
    "python-dotenv>=1.0.1",
    "waveshare-epaper>=1.3.0",
]

[tool.uv.sources]
imagen = { path = "../imagen", editable = true }


